Adding support for executing pre-defined SQL queries with parameters.
This commit adds
 - the Query model
 - the QueryController
 - the views/query/* views.

Also, this commit adds an early version of the SQL queries, developed by Vaggelis.
The queries are located in the directory protected/data/queries/
Query 1 and 3 work fine. The rest queries need the parameter names to be added.

The views/query/* views also the prettify extension to present the SQL code nicely.
Also, the SDataProvider was added, which creates an HTML table from a PHP array variable.